generics(PreProcess)	R Documentation

OOMPA Generic Functions

Description

New generic functions for processing and analyzing microarrays.

Usage


process(object, action, parameter = NULL)
analyze(object, ...)

Arguments

object 	 Any OOMPA class representing a microarrays or a set of microarrays

action 	 The action to process the class.

parameter 	 Any parameters needed to execute the process.

... 	 Place holder for additional parameters needed in derived classes

Details

In general, the analyze method represents an expensive computational step carried out in preparation for a graphical display, but the semantics may differ from class to class. The default implementation of the method performs the null analysis; that is, the return value is identical to the object that is passed in as the first argument.

The process method represents a function that acts on the data of some object to process it in some way. For example, normalizing a set of microarray data is typically one processing step in a long series that is required to take the raw data and turn it into something useful.

Value

Varies depending on the implementation in derived classes. Typically another object of the same or a closely related class.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Processor, Pipeline

Examples


# these are generic functions...

[Package PreProcess version 2.7.0 Index]