graph.utility(PreProcess)	R Documentation

OOMPA graphical utility functions

Description

Utility functions for graphics.

Usage


ellipse(a, b, x0=0, y0=0, ...)
f.qq(x, main = "", cut = 0, ...)
f.qt(x, df, main = "", cut = 0, ...)

Arguments

a 	 Half the length of the elliptical axis in the x-direction

b 	 Half the length of the elliptical axis in the y-direction

x0 	 X-coordinate of the center of the ellipse

y0 	 Y-coordinate of the center of the ellipse

main 	 A text string

cut 	 A real number

df 	 An integer; the number of degrees of freedom in the t-test

... 	 Additional graphical parameters passed on to lower-level functions

x 	 A numeric vector

Details

The ellipse function draws an ellipse on an existing plots. The ellipses produced by this function are oriented with their major and minor axes parallel to the coordinate axes. The current implementation uses points internally.

The function f.qq is a wrapper that combines qqnorm and qqline into a single function call.

The function f.qt is a wrapper that produces quantile-quantile plots comparing the observed vector x with a T-distribution.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

See also points

Examples


x <- rnorm(1000, 1, 2)
y <- rnorm(1000, 1, 2)
plot(x,y)
ellipse(1, 1, col=6, type='l', lwd=2)
ellipse(3, 2, col=6, type='l', lwd=2)
f.qq(x, main='Demo', col='blue')
f.qq(x, cut = 3)
f.qt(x, df = 3)
f.qt(x, df = 40)

[Package PreProcess version 2.7.0 Index]