matrix.utility(PreProcess)	R Documentation

OOMPA Matrix Utility Functions

Description

Utility functions for manipulating matrices.

Usage


flipud(x)
fliplr(x)

Arguments

x 	 a matrix

Value

The flipud function returns a matrix the same size as x, with the order of the rows reversed, so the matrix has been flipped vertically. The fliplr function returns a matrix the same size as x but flipped horizontally, with the order of the columns reversed.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

Examples


mat <- matrix(1:6, 2, 3)
mat
flipud(mat)
fliplr(mat)

[Package PreProcess version 2.7.0 Index]