stat.utility(PreProcess)	R Documentation

OOMPA Statistical Utility Functions

Description

Utility functions for statistical computations.

Usage


f.above.thresh(a, t)
f.cord(x, y, inf.rm)
f.oneway.rankings(r, s)
stdize(x, center=TRUE, scale=TRUE)

Arguments

a 	 a vector

t 	 a real number

x 	 a vector

y 	 a vector

inf.rm 	 a logical value

r

s

center

scale

Value

f.above.thresh returns the fraction of elements in the vector a that are greater than the threshold t.
f.cord returns the concordance coefficient between the two input vectors x and y. If inf.rm is true, then infinite values are removed before computing the concordance; missing values are always removed.
stdize is a synonym for scale.
f.oneway.rankings is implemented as order(s)[r] and I cannot recall why we defined it or where we used it.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

Examples


x <- rnorm(1000, 1, 2)
y <- rnorm(1000, 1, 2)
f.above.thresh(x, 0)
f.above.thresh(y, 0)
f.cord(x, y)
z <- stdize(x)
f.cord(x, z)

[Package PreProcess version 2.7.0 Index]