semantic colors(PreProcess)	R Documentation

Pre-defined colors for microarray plots

Description

A collection of predefined color names to help ensure consistency in multiple graphical displays of microarray data.

COLOR.BORING:

Used to mark uninteresting points in a plot; gray.

COLOR.SIGNIFICANT:

Used to mark points that are statistically significant; red

COLOR.EXPECTED:

Used to draw curves representing an expected distribution; blue

COLOR.OBSERVED:

Used to draw curves indicating the observed distribution; darkgreen

COLOR.PERMTEST:

Used to draw curves indicating distributions derived from a permutation test; magenta

COLOR.FITTED:

Used to draw curves obtained by some fitting procedure, such as loess; orange

COLOR.CENTRAL.LINE:

Used to draw lines through the centers of distributions or expected values; blue

COLOR.CONFIDENCE.CURVE:

Used to draw confidence bounds around curves; red3

COLOR.BAD.REPLICATE:

Used to indicate highly variable points; purple1

COLOR.WORST.REPLICATE:

Used to mark extraordinarily variable points; purple3

COLOR.FOLD.DIFFERENCE:

Used to indicate points with large fold difference; skyblue

COLOR.BAD.REPLICATE.RATIO:

Used to flag points for which the ratios at replicate spots are highly variable; violetred

COLOR.TOP.TEN:

Used to mark points in the "top ten" list; cadetblue

COLOR.BOTTOM.TEN:

Used to mark points in "bottom ten" list of most underexpressed genes; pink

COLOR.BOTTOM.TEN.SOLO:

Use unknown; palegreen

COLOR.TOP.TEN.SOLO:

Use unknown; deeppink

Examples


x <- seq(0, 2*pi, by=0.1)
plot(x, sin(x), col=COLOR.BORING)

[Package PreProcess version 2.7.0 Index]