ChannelType-class(PreProcess)	R Documentation

The ChannelType class

Description

This class represents the "type" of a microarray channel.

Usage


ChannelType(mk, md, nc, nr, gl, design = "")
setDesign(object, design)
getDesign(object)
## S4 method for signature 'ChannelType':
print(x,...)
## S4 method for signature 'ChannelType':
summary(object,...)

Arguments

mk 	 A string; the name of the manufacturer of the microarray (e.g., 'Affymetrix')

md 	 A string; the model of the microarray (e.g., 'Hu95A')

nc 	 An integer; the number of columns in the array

nr 	 An integer; the number of rows in the array

gl 	 A string; the material used to label samples

design 	 A string containing the name of an object describing details about the design of the microarray

object 	 A ChannelType object

x 	 A ChannelType object

... 	 Additional arguments are as in the underlying generic methods.

Details

Microarrays come in numerous flavors. At present, the two most common types are the synthesized oligonucleotide arrays produced by Affymetrix and the printed cDNA arrays on glass, which started in Pat Brown's lab at Stanford. In earlier days, it was also common to find nylon microarrays, with the samples labeled using a radiactive isotope. The glass arrays are distinguished from other kinds of arrays in that they typically cohybridize two different samples imultaneously, using two different fluorescent dyes. The fluorescence from each dye is scanned separately, producing two images and thus two related sets of data from the same microarray. We refer to these parallel data sets within an array as ``channels''.

An object of the ChannelType class represents a combination of the kind of microarray along with the kind of labeling procedure. These objects are intended to be passed around as part of more complex objects representing the actual gene expression data collected from particular experiments, in order to be able to eventually tie back into the description of what spots were laid down when the array was produced.

The ChannelType object only contains a high level description of the microarray, however. Detailed information about what biological material was laid down at each spot on the microarray is stored elsewhere, in a ``design'' object. Within a ChannelType object, the design is represented simply by a character string. This string should be the name of a separate object containing the detailed design information. This implementation allows us to defer the design details until later. It also saves space by putting the details in a single object instead of copying them into every microarry. Finally, it allows that single object to be updated when better biological annotations are available, with the benefits spreading immediately to all the microarray projects that use that design.

Value

The ChannelType constructor reutrns a valid object of the class.
The setDesign function invisibly returns the ChannelType object on which it was invoked.
The getDesign function returns the design object referred to by the design slot in the ChannelType object. If this string does not evaluate to the name of an object, then getDesign returns a NULL value.

Slots

maker:

A string; the name of the manufacturer of the microarray (e.g., 'Affymetrix')

model:

A string; the model of the microarray (e.g., 'Hu95A')

nCol:

An integer; the number of columns in the array

nRow:

An integer; the number of rows in the array

glow:

A string; the material used to label samples

design:

A string containing the name of an object describing details about the design of the microarray

Methods

print(x, ...)

Prints all the information in the object

summary(object, ...)

Writes out a summary of the object

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Channel

Examples


x <- ChannelType('Affymetrix', 'oligo', 100, 100, 'fluor')
print(x)

summary(x)

y <- setDesign(x, 'fake.design')
print(y)
summary(y)
d <- getDesign(y)
d

rm(d, x, y) # cleanup

[Package PreProcess version 2.7.0 Index]