Channel-class(PreProcess)	R Documentation

The Channel class

Description

An object of the Channel class represents a single kind of measurement performed at all spots of a microarray channel. These objects are essentially just vectors of data, with length equal to the number of spots on the microarray, with some extra metadata attached.

Usage


Channel(parent, name, type, x)
## S4 method for signature 'Channel, missing':
plot(object, ...)
## S4 method for signature 'Channel':
hist(object, ...)
## S4 method for signature 'Channel':
summary(object, ...)
## S4 method for signature 'Channel':
print(object, ...)
## S4 method for signature 'Channel':
image(object, ...)

Arguments

parent 	 A string representing the name of a parent object from which this object was derived

name 	 A string with a displayable name for this object

type 	 A ChannelType object

x 	 A numeric vector

object 	 A Channel object

... 	 Additional arguments are as in the underlying generic methods.

Details

As described in the help pages for ChannelType, each microarray hybridization experiment produces one or more channels of data. Channel objects represent a single measurement performed at spots in one microarray channel. The raw data from a full experiment typically contains multiple measurements in mutliple channels.

The full set of measurements is often highly processed (by, for example, background subtraction, normalization, log transformation, etc.) before it becomes useful. We have added a history slot that keeps track of how a Channel was produced. By allowing each object to maintain a record of its history, it becomes easier to document the processing when writing up the methods for reports or papers. The history slot of the object is updated using the generic function process together with a Processor object.

Value

The print, hist, and image methods all invisibly return the Channel object on which they were invoked. The print and summary methods return nothing.

Slots

parent:

A string representing the name of a parent object from which this object was derived.

name:

A string with a displayable name for this object

type:

A ChannelType object

x:

A numeric vector

history:

A list that keeps a record of the calls used to produce this object

Methods

print(object, ...)

Print all the data on the object. Since this includes the entire data vector, you rarely want to do this.

summary(object, ...)

Write out a summary of the object.

plot(object, ...)

Produce a scatter plot of the measurement values in the slot x of the object against their index , which serves as a surrogate for the position on the microarray. Additional graphical parameters are passed along.

hist(object, ...)

Produce a histogram of the data values in slot x of the object. Additional graphical parameters are passed along.

image(object, ...)

This method produces a two-dimensional "cartoon" image of the measurement values, with the position in the cartoon corresponding to the two-dimensional arrangement of spots on the actual microarray. Additional graphical parameters are passed along.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

ChannelType, process, Processor

Examples


# simulate a moderately realistic looking microarray
nc <- 100                       # number of rows
nr <- 100                       # number of columns
v <- rexp(nc*nr, 1/1000)        # "true" signal intensity (vol)
b <- rnorm(nc*nr, 80, 10)       # background noise
s <- sapply(v-b, max, 1)        # corrected signal instensity (svol)
ct <- ChannelType('user', 'random', nc, nr,  'fake')
raw <- Channel(name='fraud', type=ct, parent='', x=v)
subbed <- Channel(name='fraud', parent='', type=ct, x=s)
rm(nc, nr, v, b, s)             # clean some stuff

summary(subbed)
summary(raw)

par(mfrow=c(2,1))
plot(raw)
hist(raw)

par(mfrow=c(1,1))
image(raw)

# finish the cleanup
rm(ct, raw, subbed)

[Package PreProcess version 2.7.0 Index]