CompleteChannel-class(PreProcess)	R Documentation

The CompleteChannel class

Description

An object of the CompleteChannel class represents one channel (red or green) of a two-color fluorescence microarray experiment. Alternatively, it can also represent the entirety of a radioactive microarray experiment. Affymetrix experiments produce data with a somewhat different structure because they use multiple probes for each target gene.

Usage


CompleteChannel(name, type, data)
## S4 method for signature 'CompleteChannel':
print(x, ...)
## S4 method for signature 'CompleteChannel':
summary(object, ...)
## S4 method for signature 'CompleteChannel':
as.data.frame(x, row.names=NULL, optional=FALSE)
## S4 method for signature 'CompleteChannel, missing':
plot(x, useLog=FALSE, ...)
## S4 method for signature 'CompleteChannel':
image(x, ...)
## S4 method for signature 'CompleteChannel':
analyze(object, useLog=FALSE, ...)
## S4 method for signature 'CompleteChannel, Processor':
process(object, action, parameter)
## S4 method for signature 'CompleteChannel':
channelize(object)

Arguments

name 	 A string containing the name of the object

type 	 A ChannelType object

data 	 A data frame. For the pre-defined ``extraction'' processors to work correctly, this should include columns called vol, bkgd, svol, SD, and SN.

x 	 A CompleteChannel object

object 	 A CompleteChannel object

useLog 	 A logical value

action 	 A Processor object used to process a CompleteChannel.

parameter 	 Any object that makes sense as a parameter to the function represented by the Processor action

row.names 	 See as.data.frame

optional 	 See as.data.frame

... 	 Additional arguments are as in the underlying generic methods.

Details

The names come from the default column names in the ArrayVision software package used at M.D. Anderson for quantifying glass or nylon microarrays. Column names used by other software packages should be mapped to these.

Value

The analyze method returns a list of three density functions.
The return value of the process function depends on the Processor performing the action, but is typically a Channel object.
Graphical methods invisibly return the object on which they were invoked.

Slots

name:

A string containing the name of the object

type:

A ChannelType object

data:

A data frame

history:

A list that keeps a record of the calls used to produce this object

Methods

print(x, ...)

Print all the data on the object. Since this includes the data frame, you rarely want to do this.

summary(object, ...)

Write out a summary of the object.

as.data.frame(x,row.names=NULL, optional=FALSE)

Convert the CompleteChannel object into a data frame. As you might expect, this simply returns the data frame in the data slot of the object.

plot(x, useLog=FALSE, ...)

Produces three estimated density plots: one for the signal, one for the background, and one for the background-corrected signal. Additional graphical parameters are passed along. The logical flag useLog determines whether the data are log-transformed before estimating and plotting densities.

analyze(object, useLog=FALSE, ...)

This method computes the estimated probability density functions for the three data components (signal, background, and background-corrected signal), and returns them as a list.

image(object, ...)

Uses the image method for Channel objects to produce geographically aligned images of the log-transformed intensity and background estimates.

channelize(object)

A string giving the name of the class of a channel that is produced when you process a CompleteChannel object.

process(object, action, parameter=NULL)

Use the Processor action to process the CompleteChannel object. Returns an object of the class described by channelize, which defaults to Channel.

Pre-defined Processors

The library comes with several Processor objects already defined; each one takes a CompleteChannel as input, extracts a single value per spot, and produces a Channel as output.

PROC.BACKGROUND

Extract the vector of local background measurements.

PROC.SIGNAL

Extract the vector of foreground signal intensity measurements.

PROC.CORRECTED.SIGNAL

Extract the vector of background-corrected signal measurements. Note that many software packages automatically truncate these value below at zero, so this need not be the same as SIGNAL - BACKGROUND.

PROC.NEG.CORRECTED.SIGNAL

Extract the vector of background-corrected signal intensities by subtracting the local background from the observed foreground, without truncation.

PROC.SD.SIGNAL

Extract the vector of pixel standard deviations of the signal intensity.

PROC.SIGNAL.TO.NOISE

Extract the vector of signal-to-noise ratios, defined as CORRECTED.SIGNAL divided by the standard deviation of the background pixels.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

process, Processor, Pipeline, Channel, as.data.frame

Examples


# simulate a complete channel object

v <- rexp(10000, 1/1000)
b <- rnorm(10000, 60, 6)
s <- sapply(v-b, function(x) {max(0, x)})
ct <- ChannelType('user', 'random', 100, 100, 'fake')
x <- CompleteChannel(name='fraud', type=ct,
                      data=data.frame(vol=v, bkgd=b, svol=s))
rm(v, b, s, ct)

summary(x)

opar <- par(mfrow=c(2,3))
plot(x)
plot(x, main='Log Scale', useLog=TRUE)
par(opar)

opar <- par(mfrow=c(2,1))
image(x)
par(opar)

b <- process(x, PROC.NEG.CORRECTED.SIGNAL)
summary(b)

q <- process(b, PIPELINE.STANDARD)
summary(q)

q <- process(x, PIPELINE.MDACC.DEFAULT)
summary(q)

# cleanup
rm(x, b, q, opar)

[Package PreProcess version 2.7.0 Index]