channelize(PreProcess)	R Documentation

Object-oriented conversion from CompleteChannel to Channel

Description

Provide a generic function for propagating the class of derived objects through a processing pipeline.

Usage


channelize(object)

Arguments

object 	 An object of a clas derived from CompleteChannel.

Details

Having abstracted away the notion of extracting a particular measurement from a CompleteChannel object and producing a simple Channel, we need a way to allow object-oriented programming and derived classes to work with our Processor and Pipeline routines. The underlying idea is that specific kinds of microarrays or specific software to quantify microarrays might have special properties that should be exploited in processing. For example, the first few generations of miroarrays printed at M.D. Anderson spotted every cDNA clone in duplicate. The analysis of such arrays should exploit this additional structure. In order to do so, we must derive classes from CompleteChannel and Channel and ensure that the classes of extracted objects are propagated correctly through the processing pipeline. The channelize methods achieves this goal.

Value

Returns a string, which represents the name of a class (suitable for passing to the new constructor) extracted from an object belonging to a class derived from CompleteChannel.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Processor, Pipeline, Channel, CompleteChannel

[Package PreProcess version 2.7.0 Index]