BumSummary-class(ClassComparison)	R Documentation

The BumSummary class

Description

An implementation class. Users are not expected to create these objects directly; they are produced as return objects from the summary method for Bum.

Slots

bum:

Object of class Bum

estimates:

Object of class data.frame

Fhat:

Object of class numeric

Methods

show

signature(object = "BumSummary"): Print the object, which contains a summary of the underlying Bum object. The summary contains a data frame with estimates of the fraction of true positives (TP), false postives (FP), true negatives (TN) and false negatives (FN) at the set of p-value cutoffs specified in the call to the summary method.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Bum

Examples


# See the examples for the Bum class.

[Package ClassComparison version 2.5.0 Index]