significant(ClassComparison)	R Documentation

Generic Functions for Significance

Description

In the world of multiple testing that is inhabited by most microarray or protein profiling experiments, analysts frequently perform separate statistical tests for each gene or protein in the experiment. Determining cutoffs that achieve statistical significance (in a meaningful way) is an inherent part of the procedure. It is then common to select the significant items for further processing or for preparing reports, or at least to count the number of significant items. These generic functions provide a standard set of tools for selecting and counting the significant items, which can be used with various statistical tests and various ways to account for multiple testing.

Usage


cutoffSignificant(object, ...)
selectSignificant(object, ...)
countSignificant(object, ...)

Arguments

object 	 Typically, an object that performs multiple statistical tests on microarray or proteomics data.

... 	 All generic methods are designed to take optional additional arguments for flexibility in creating derived classes.

Value

cutoffSignificant returns appropriate cutoff values that achieve specified signficance criteria.
selectSignificant returns a logical vector, with true values indicating items that satisfy the cutoff making them statistically significant.
countSignificant returns an integer, representing the number of significant items.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Classes that implement these methods include Bum, Sam, Dudoit, MultiWilcoxonTest, and TNoM.

Examples


# Since these are generic methods, there are no examples. See the
# classes that implement specific instances.

[Package ClassComparison version 2.8.0 Index]