MultiTtest-class(ClassComparison)	R Documentation

The MultiTtest Class

Description

Class to perform row-by-row t-tests on microarray or proteomics data.

Usage


MultiTtest(data, classes)
## S4 method for signature 'MultiTtest':
summary(object, ...)
## S4 method for signature 'MultiTtest':
hist(x, xlab='T Statistics', main=NULL, ...)
## S4 method for signature 'MultiTtest, missing':
plot(x, ylab='T statistics', ...)
## S4 method for signature 'MultiTtest, ANY':
plot(x, y, xlab='T Statistics', ylab=deparse(substitute(y)), ...)

Arguments

data 	 Either a data frame or matrix with numeric values or an ExpressionSet as defined in the BioConductor tools for analyzing microarray data.

classes 	 If data is a data frame or matrix, then classes must be either a logical vector or a factor. If data is an ExpressionSet, then classes can be a character string that names one of the factor columns in the associated phenoData subobject.

object 	 A MultiTtest object

x 	 A MultiTtest object

y 	 A numeric vector

xlab 	 Label for the x axis

ylab 	 Label for the y axis

main 	 Plot title

... 	 The usual graphical parameters can by supplied to hist and plot methods.

Value

The graphical routines invisibly return the object against which they were invoked.

Creating objects

Although objects can be created using new, the better method is to use the MultiTtest function. In the simplest case, you simply pass in a data matrix and a logical vector assigning classes to the columns, and the constructor performs row-by-row two-sample t-tests and computes the associated (single test) p-values. To adjust for multiple testing, you can pass the p-values on to the Bum class.

If you use a factor instead of a logical vector, then the t-test compares the first level of trhe factor to everything else. To handle the case of multiple classes, see the MultiLinearModel class.

As with other class comparison functions that are part of the OOMPA, we can also perform statistical tests on ExpressionSet objects from the BioConductor libraries. In this case, we pass in an ExpressionSet object along with the name of a factor to use for splitting the data.

Slots

t.statistics:

Object of class numeric containing the computed t-statistics.

p.values:

Object of class numeric containing the computed p-values.

groups:

Object of class character containing the names of the classes being compared.

call:

Object of class call containing the function call that created the object.

Methods

summary(object, ...)

Write out a summary of the object.

hist(x, xlab='T Statistics', main=NULL, ...)

Produce a histogram of the t-statistics.

plot(x)

Produces a scatter plot of the t-statistics against their index.

plot(x,y)

Produces a scatter plot of the t-statistics in the object x against the numeric vector y.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

matrixT, Bum, MultiLinearModel, Dudoit

Examples


ng <- 10000
ns <- 50
dat <- matrix(rnorm(ng*ns), ncol=ns)
cla <- factor(rep(c('A', 'B'), each=25))
res <- MultiTtest(dat, cla)
summary(res)
hist(res, breaks=101)
plot(res)
plot(res, res@p.values)
hist(res@p.values, breaks=101)

rm(ng, ns, dat, cla, res)

[Package ClassComparison version 2.5.0 Index]