variantT(ClassComparison)	R Documentation

Classes for Variant T-tests

Description

Class to perform row-by-row paired or unequal variance t-tests on microarray or proteomics data.

Usage


MultiTtestPaired(data, classes, pairing)
MultiTtestUnequal(data, classes)
## S4 method for signature 'MultiTtestPaired':
summary(object, ...)
## S4 method for signature 'MultiTtestUnequal':
summary(object, ...)

Arguments

data 	 Either a data frame or matrix with numeric values or an ExpressionSet as defined in the BioConductor tools for analyzing microarray data.

classes 	 If data is a data frame or matrix, then classes must be either a logical vector or a factor. If data is an ExpressionSet, then classes can be a character string that names one of the factor columns in the associated phenoData subobject.

pairing 	 A numerical vector indicating which samples are paired.

object 	 A MultiTtest object

... 	 Unused; optional extra parameters for summary.

Creating objects

Although objects can be created using new, the better method is to use the MultiTtestPaired or MultiTtestUnequal functions. In the simplest case, you simply pass in a data matrix and a logical vector assigning classes to the columns (and, in the case of a paired t-test, a numeric vector describing the pairing), and the constructor performs row-by-row two-sample t-tests and computes the associated (single test) p-values. To adjust for multiple testing, you can pass the p-values on to the Bum class.

If you use a factor instead of a logical vector, then the t-test compares the first level of the factor to everything else. To handle the case of multiple classes, see the MultiLinearModel class.

As with other class comparison functions that are part of the OOMPA, we can also perform statistical tests on ExpressionSet objects from the BioConductor libraries. In this case, we pass in an ExpressionSet object along with the name of a factor to use for splitting the data.

Extends

Both classes extend class MultiTtest, directly. See that class for descriptions of the inherited methods and slots.

Slots

df:

The MultiTtestUnequal class adds a slot to record e gene-by-gene degrees of freedom, which can change along with the variances.

Methods

summary

signature(object = MultiTtestPaired): Write out a summary of the object.

summary

signature(object = MultiTtestUnequal): Write out a summary of the object.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

References

None.

See Also

MultiTtest, Bum

Examples


ng <- 10000
ns <- 50
dat <- matrix(rnorm(ng*ns), ncol=ns)
cla <- factor(rep(c('A', 'B'), each=25))
res <- MultiTtestUnequal(dat, cla)
summary(res)
hist(res, breaks=101)
plot(res, res@p.values)

pairing <- rep(1:25, 2)
res <- MultiTtestPaired(dat, cla, pairing)
summary(res)
plot(res)
hist(res@p.values, breaks=101)

rm(ng, ns, dat, cla, res, pairing)

[Package ClassComparison version 2.8.0 Index]