matrixT(ClassComparison)	R Documentation

Multiple T Tests by Matrix Multiplication

Description

Utility functions for computing vectors of row-by-row means, variances, and t-statistics.

Usage


matrixMean(x)
matrixUnequalT(m, v)
matrixPairedT(m, v, pf)
matrixVar(x, xmean)
matrixT(m, v)

Arguments

x 	 a matrix

xmean 	 a numeric vector or single-column matrix

m 	 a matrix

v 	 a logical vector of length equal to the number of columns of m

pf 	 a numerical vector of length equal to the number of columns of m, indicating which samples should be paired

Value

matrixMean returns a single-column matrix containing the row-by-row means of x.
matrixVar returns a single-column matrix containing the row-by-row means of x, assuming that xmean contains the corresponding mean values.
matrixT returns a single-column matrix of t-statistics from a two-sample t-test comparing the columns for which v is true to those for which v is false.
matrixPairedT returns a single-column matrix of t-statistics from a paired t-test.
matrixUnequalT returns a list with two components: tt is a single-column matrix of t-statistics from a two-sample unequal variance t-test comparing the columns for which v is true to those for which v is false, and df is a single-column matrix of the degrees of freedom associated with each row..

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

MultiTtest

Examples


ng <- 1000
ns <- 50
dat <- matrix(rnorm(ng*ns), ncol=ns)
clas <- factor(rep(c('Good', 'Bad'), each=25))
myMean <- matrixMean(dat)
myVar  <- matrixVar(dat, myMean)
plot(myMean, myVar)

myT <- matrixT(dat, clas)
hist(myT)

rm(ng, ns, dat, myMean, myVar, myT)

[Package ClassComparison version 2.9.0 Index]