dwil(ClassComparison)	R Documentation

Wilcoxon Density Function

Description

Computes the density function for the Wilcoxon rank-sum distribution without centering.

Usage


dwil(q, m, n)

Arguments

q 	 vector of quantiles

m 	 number of observations in the first sample

n 	 number of observations in the second sample

Details

Computes the density function for the Wilcoxon rank-sum distribution, using exact values when both groups have fewer than 50 items and switching to a normal approximation otherwise. It was originally written for S-Plus, which still perversely insists that m and n must be less than 50. The function was retained when the OOMPA library was ported to R, since S-Plus keeps the actual rank-sum but R centers the distribution at zero. This function encapsulated the difference, allowing everything else to continue to work as it had worked previously.

Value

A vector of the same length as q containing (approximate or exact) values of the density function.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

MultiWilcoxonTest

Examples


dwil(51:60, 9, 3)
dwil(51:60, 9, 51)

[Package ClassComparison version 2.8.0 Index]