TNoMSummary-class(ClassComparison)	R Documentation

The TNoMSummary Class

Description

An implementation class. Users are not expected to create these objects directly; they are produced as return objects from the summary method for TNoM.

Slots

TNoM:

Object of class TNoM ~~

counts:

Object of class numeric ~~

Methods

show

signature(object = TNoMSummary): Print the object, which contains a summary of the underlying TNoM object. In particular, the summary reports the number of genes acheiving each possible number of misclassifications.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

TNoM

Examples


# See the examples for the TNoM class.

[Package ClassComparison version 2.5.0 Index]