SamSummary-class(ClassComparison)	R Documentation

The SamSummary Class

Description

An implementation class. Users are not expected to create these objects directly; they are produced as return objects from the summary method for Sam.

Slots

fdr:

A number between 0 and 1; the expected false discovery rate

hi:

Upper threshold for significance

lo:

Lower threshold for significance

cutoff:

A numeric value specified in the call to the Sam summary method.

significant.calls:

vector of logical values

average.false.count:

The average number of false positives in the permuted data at this cutoff level.

Methods

show

signature(object = SamSummary): Print the object, which contains a summary of the underlying Sam object.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Sam

Examples


# See the examples for the Sam class.

[Package ClassComparison version 2.5.0 Index]