SingleGroup-class(ClassComparison)	R Documentation

The SingleGroup Class

Description

Preliminary analysis of one group of samples for use in the SmoothTtest class. A key feature is the standard quality control plot.

Usage


SingleGroup(avg, sd, span = 0.5, name = '')
## S4 method for signature 'SingleGroup':
as.data.frame(x, row.names=NULL, optional=FALSE)
## S4 method for signature 'SingleGroup':
summary(object, ...)
## S4 method for signature 'SingleGroup':
print(x, ...)
## S4 method for signature 'SingleGroup, missing':
plot(x, multiple=3, ccl=0, main=x@name,
 xlab='Mean', ylab='Std Dev', xlim=0, ylim=0, ...)

Arguments

avg 	 A numeric vector of mean values

sd 	 A numeric vector of standard deviations

span 	 The span parameter is passed onto loess.

name 	 A character string; the name of this object

object 	 A SingleGroup object

x 	 A SingleGroup object

multiple 	 A real number; the multiple of the smoothed standard deviation to call significant.

ccl 	 A list containing objects of the ColorCoding class. If left at its default value of zero, colors are chosen automatically.

main 	 Plot title

xlab 	 Label for the x axis

ylab 	 Label for the y axis

xlim 	 Plotting limits for the x axis. If left at the default value of zero, then the limits are automatically generated

ylim 	 Plotting limits for the y axis. If left at the default value of zero, then the limits are automatically generated

row.names 	 See the base version of as.data.frame.default

optional 	 See the base version of as.data.frame.default

... 	 The usual extra parameters to generic or plotting routines

Details

In 2001 and 2002, Baggerly and Coombes developed the smooth t-test for finding differentially expressed genes in microarray data. Along with many others, they began by log-transforming the data as a reasonable step in the direction of variance stabilization. They observed, however, that the gene-by-gene standard deviations still seemed to vary in a systematic way as a function of the mean log intensity. By borrowing strenght across genes and using loess to fit the observed standard deviations as a function of the mean, one presumably got a better estimate of the true standard deviation.

Creating Objects

Objects can be created by calls to the SingleGroup constructor. Users rarely have need to create these objects directly; they are usually created as a consequence of the construction of an object of the SmoothTtest class.

Slots

name:

A character string; the name of this object

avg:

The numeric vector of mean values

sd:

The numeric vector of standard deviations

span:

The span parameter used in the loess function to fit sd as a function of avg.

fit:

A list containing components x and y resulting from the loess fit.

score:

A numeric vector; the ratio of the pointwise standard deviations to their smooth (loess) estimates.

Methods

as.data.frame(x, row.names=NULL, optional=FALSE)

Combine the slots containing numeric vectors into a data frame, suitable for printing or exporting.

summary(object, ...)

Write out a summary of the object.

print(x, ...)

Print the entire object. You never want to do this.

plot(x, multiple=3, ccl=0, main=x@name, xlab='Mean', ylab='Std Dev', xlim=0, ylim=0, ...)

Produce a scatter plot of the standard deviations (x@sd) as a function of the means (x@avg). The appropriate mutliple of the loess fit is overlaid, and points that exceed this multiple are flaged in a different color. Colors in the plotare controlled by the current values of COLOR.CENTRAL.LINE, COLOR.CONFIDENCE.CURVE, COLOR.BORING, COLOR.BAD.REPLICATE, and COLOR.WORST.REPLICATE.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

References

Baggerly, K.A., Coombes, K.R., Hess, K.R., Stivers, D.N., Abruzzo, L.V., Zhang, W. Identifying differentially expressed genes in cDNA microarray experiments. J Comp Biol. 8:639-659, 2001.

Coombes, K.R., Highsmith, W.E., Krogmann, T.A., Baggerly, K.A., Stivers, D.N., Abruzzo, L.V. Identifying and quantifying sources of variation in microarray data using high-density cDNA membrane arrays. J Comp Biol. 9:655-669, 2002.

See Also

SmoothTtest

Examples


m <- rnorm(1000, 8, 2.5)
v <- rnorm(1000, 0.7)
plot(m, v)

x <- SingleGroup(m, v, name='bogus')

summary(x)

plot(x)
plot(x, multiple=2)

# cleanup
rm(m, v, x)

[Package ClassComparison version 2.9.0 Index]