numeric or NULL-class(oompaBase)	R Documentation

Class 'numeric or NULL'

Description

This class is a union that can represent either a numeric vector or a NULL value.

Objects from the Class

A virtual Class: No objects may be created from it.

Methods

No methods defined with class "numeric or NULL" in the signature.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

Processor uses the numeric or NULL class to represent the default values of its parameters.

[Package oompaBase version 2.10.0 Index]