TailRankTest-methods(TailRank)	R Documentation

Methods for TailRankTest objects

Description

This file describes the methods for an object of the class TailRankTest class.

Usage


## S4 method for signature 'TailRankTest':
summary(object, ...)
## S4 method for signature 'TailRankTest':
hist(x, overlay, ...)
## S4 method for signature 'TailRankTest':
as.logical(x, ...)
## S4 method for signature 'TailRankTest':
getStatistic(object,...)

Arguments

x 	 A TailRankTest object

object 	 A TailRankTest object

overlay 	 An optional logical flag; defaults to FALSE.

... 	 Extra graphical parameters

Value

as.logical 	 Returns a logical vector. TRUE values pick out candidate biomarkers where the tail-rank test statistic exceeds the significance cutoff.

getStatistic 	 Returns the vector of tail-rank statistics contained in object.

hist 	 Invisibly returns the TailRankTest object.

summary 	 Invisibly returns the TailRankTest object.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

TailRankTest-class, TailRankTest, tailRankPower

Examples


# generate some fake data to use in the example
nr <- 40000
nc <- 110
fake.data <- matrix(rnorm(nr*nc), ncol=nc)
fake.class <- rep(c(TRUE, FALSE), c(40, 70))

# build an object
null.tr <-  TailRankTest(fake.data, fake.class)

# summarize the object
summary(null.tr)

# plot a histogram
hist(null.tr)
hist(null.tr, breaks=70, col='blue', overlay=TRUE)

# get a logical vector that can select those markers
# identified by the test
selector <- as.logical(null.tr)

[Package TailRank version 2.8.0 Index]