biomarkerPowerTable(TailRank)	R Documentation

Power tables for the tail-rank test

Description

Compute an array of power tables for the tail-rank.test.

Usage


biomarkerPowerTable(G, N1=20, N2=seq(25, 250, by=25), psi = c(0.95, 0.99), conf=0.99, phi = seq(0.10, 0.50, by = 0.05), model="bb")

Arguments

G 	 An integer; the number of genes being assessed as potential biomarkers. Statistically, the number of hypotheses being tested.

N1 	 An integer; the number of "train" or "healthy" samples used.

N2 	 An integer; the number of "test" or "cancer" samples used.

psi 	 A real number between 0 and 1; the desired specificity of the test.

conf 	 A real number between 0 and 1; the confidence level of the results. Can be obtained by subtracting the family-wise Type I error from 1.

phi 	 A real number between 0 and 1; the sensitivity that one would like to be able to detect, conditional on the specificity.

model 	 A character string that determines whether power and significance are computed from abinomial or a beta-binomial (bb) model.

Value

Returns a list of objects of the BMPT class. Each item in the list consists of a two-dimensional table (indexed by the sample sizes N and the sensitivities phi) with scalars recording the values of G, conf, and psi that were used to generate it.

Note

Default values of the optional arguments (N, psi, conf, phi)are included in the usage examples.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

TailRankTest, tailRankPower, biomarkerPowerTable, matrixMean, toleranceBound

Examples


stuff <- biomarkerPowerTable(10000, 20,
                               c(10, 20, 50, 100, 250, 500),
                               c(0.95, 0.99),
                               c(0.99, 0.95),
                               seq(0.1, 0.7, by=0.1))
lapply(stuff, summary)

[Package TailRank version 2.8.0 Index]