BMPT-class(TailRank)	R Documentation

The BMPT Class

Description

A class for producing BioMarker Power Tables (BMPT), and methods for accessing them. This class is primarily an implementation detail for the function biomarkerPowerTable.

Usage


BMPT(G, psi, conf, power)
## S4 method for signature 'BMPT':
print(x,...)
## S4 method for signature 'BMPT':
summary(object,...)

Arguments

G 	 A positive integer.

psi 	 A real number between 0 and 1.

conf 	 A real number between 0 and 1.

power 	 A data frame.

x 	 A BMPT object.

object 	 A BMPT object.

... 	 Extra graphical parameters

Creating objects

Although objects can be created using new, the preferred method is to use the constructor function BMPT. In practice, these objects are most likely to be created using the more general interface through biomarkerPowerTable.

Slots

G:

A positive integer; the number of genes being assessed as potential biomarkers. Statistically, the number of hypotheses being tested.

psi:

A real number between 0 and 1; the desired specificity of the test.

conf:

A real number between 0 and 1; the confidence level of the results. Can be obtained by subtracting the family-wise Type I error from 1.

power:

A data frame containing the power computations. The rows are indexed by the sample size and the columns by the sensitivity.

Methods

print(x, ...)

Print the power table x.

summary(object, ...)

Summarize the power table object.

Note

See biomarkerPowerTable for examples.

Author(s)

Kevin R. Coombes <kcoombes@mdanderson.org>

See Also

TailRankTest, tailRankPower, biomarkerPowerTable

[Package TailRank version 2.7.0 Index]