clinical.info(TailRank)	R Documentation

Experimental info for the prostate cancer data set

Description

This data set provides experimental and clinical information about the (partial) prostate cancer data set included for demonstration purposes as part of the tail.rank.test package. The experiments were two-color glass microarrays printed at Stanford.

Usage

data(clinical.info)

Format

A data frame with 112 observations on the following 6 variables.

Arrays

A factor containing the barcode of the microarray on which the experiment was performed. Each of the 112 entries should be distinct.

Reference

A factor describing the reference sample used in each experiment. This was a common reference, so the identifiers here are not meaningful.

Sample

A factor identifying the test sample in each experiment. These match the codes published in the original paper.

Status

A factor with three levels identifying normal prostate (N), prostate cancer (T), or lymph node metastasis (L).

Subgroups

A factor with five levels: I II III N O. These correspond to the groups found in the original paper using clustering.

ChipType

a factor with levels new or old. At least two different print designs of microarrays were used in this experiment; this factor identifies the design.

Source

The data was originally described in the paper by Lapointe et al., and downloaded from the Stanford Microarray Database http://genome-www5.stanford.edu/.

References

Lapointe J et al. (2004) Gene expression profiling identifies clinically relevant subtypes of prostate cancer. Proc Natl Acad Sci U S A, 101, 811–816.

See Also

expression.data, gene.info, TailRankTest

Examples


data(clinical.info)
summary(clinical.info) 

[Package TailRank version 2.8.0 Index]