gene.info(TailRank)	R Documentation

Gene information for the prostate cancer data set

Description

This data set provides information about the genes included with the (partial) prostate cancer data set as part of the tail.rank.test package.

Usage

data(gene.info)

Format

A data frame with 2000 observations on the following 6 variables.

ArrayI.Spot

a numeric vector; where is this clone spotted on the old arrays

ArrayII.Spot

a numeric vector; where is this clone spotted on the new arrays

Clone.ID

a factor; the IMAGE clone identifier

Gene.Symbol

a factor; the official gene symbol

Cluster.ID

a factor; the UniGene cluster number

Accession

a factor; the GenBanlk accession number

Source

The data was originally described in the paper by Lapointe et al., and downloaded from the Stanford Microarray Database http://genome-www5.stanford.edu/. We randomly selected 2000 of the 42129 spots to include as demonstration data here.

References

Lapointe J et al. (2004) Gene expression profiling identifies clinically relevant subtypes of prostate cancer. Proc Natl Acad Sci U S A, 101, 811–816.

See Also

clinical.info, expression.data, TailRankTest

Examples


data(gene.info)
summary(gene.info)

[Package TailRank version 2.8.0 Index]